added ProductDao and ProductDaoImpl

+++ src/main/java/entity/Boots.java
@@ -4,7 +4,7 @@ public class Boots extends Product {
     private int size;
     private boolean isNaturalSkin;
 
-    public Boots(int id, String productName, double price, double weight, String color, int productCount, int size, boolean isNaturalSkin) {
+    public Boots(Long id, String productName, double price, double weight, String color, int productCount, int size, boolean isNaturalSkin) {
         super(id, productName, price, weight, color, productCount);
         this.size = size;
         this.isNaturalSkin = isNaturalSkin;
